Refactor ElseIfCleanUp to jdt.core.manipulation (#1374)
- create new ElseIfCleanUpCore and ElseIfFixCore classes
- modify CodeStyleTabPage, CleanUpPerfTest, and plugin.xml to
  refer to new core clean-up
- fixes #1373
