+++ org.eclipse.jdt.ui/plugin.xml
@@ -7132,7 +7132,7 @@
             runAfter="org.eclipse.jdt.ui.cleanup.replace_deprecated_calls">
       </cleanUp>
       <cleanUp
-            class="org.eclipse.jdt.internal.ui.fix.ElseIfCleanUp"
+            class="org.eclipse.jdt.internal.ui.fix.ElseIfCleanUpCore"
             id="org.eclipse.jdt.ui.cleanup.else_if"
             runAfter="org.eclipse.jdt.ui.cleanup.variables">
       </cleanUp>
